requirements.txt
accelerate==0.28.0
aiofiles==23.2.1
alembic==1.13.1
altair==5.3.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
arxiv==2.1.0
asgiref==3.7.2
asttokens==2.4.1
Automat==22.10.0
backoff==2.2.1
BaiduSpider==1.0.2.6
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.100
botocore==1.34.100
Brotli==1.1.0
bs4==0.0.2
build==1.1.1
cachetools==5.3.3
cffi==1.16.0
chardet==5.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
chromedriver-autoinstaller==0.6.4
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
configobj==5.0.8
configparser==7.0.0
constantly==23.10.4
contourpy==1.2.1
courlan==1.0.0
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
dashscope==1.20.1
dataclasses-json-speakeasy==0.5.11
datasets==2.18.0
dateparser==1.2.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docker-pycreds==0.4.0
elastic-transport==8.13.0
elasticsearch==8.13.0
emoji==2.10.1
entrypoints==0.4
et-xmlfile==1.1.0
etelemetry==0.3.1
executing==2.0.1
faiss==1.7.4
faiss-cpu==1.8.0
Faker==24.9.0
fastapi==0.110.0
favicon==0.7.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.13.1
filetype==1.2.0
fire==0.6.0
FlagEmbedding==1.2.7
flatbuffers==24.3.7
fonttools==4.51.0
free-proxy==1.1.1
frontend==0.0.3
frozendict==2.4.2
fschat==0.2.36
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.19.0
google-auth==2.28.2
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
graphviz==0.20.1
greenlet==3.0.3
groq==0.5.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
htbuilder==0.6.2
html2text==2020.1.16
htmldate==1.8.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
humanize==4.9.0
hyperlink==21.0.0
importlib-metadata==6.11.0
importlib_resources==6.3.1
incremental==22.10.0
intel-openmp==2021.4.0
ipykernel==6.26.0
ipython==8.18.0
isodate==0.6.1
itemadapter==0.9.0
itemloaders==1.2.0
itsdangerous==2.2.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpath-python==1.0.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jusText==3.0.0
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.14
langchain-aws==0.1.3
langchain-community==0.0.32
langchain-google-genai==1.0.1
langchain-groq==0.1.3
langchain-openai==0.1.1
langcodes==3.3.0
langdetect==1.0.9
llama-hub==0.0.79.post1
llama-index==0.10.29
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.11
llama-index-core==0.10.29
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-huggingface==0.1.4
llama-index-llms-openai==0.1.15
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.18
llama-index-readers-llama-parse==0.1.4
llama-index-readers-web==0.1.9
llama-index-vector-stores-elasticsearch==0.1.7
llama-index-vector-stores-faiss==0.1.2
llama-parse==0.4.1
llamaindex-py-client==0.1.18
loader==2017.9.11
looseversion==1.3.0
lxml==5.1.0
Mako==1.3.3
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
merkle-json==1.0.0
millify==0.1.1
minify_html==0.15.0
mkl==2021.4.0
mkl-service==2.4.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multiprocess==0.70.16
munch==4.0.0
nest-asyncio==1.5.8
networkx==3.2.1
newspaper3k==0.2.8
nh3==0.2.17
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
oauthlib==3.2.2
ollama==0.1.8
omegaconf==2.3.0
onnxruntime==1.17.1
openai==1.14.1
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
outcome==1.3.0.post0
overrides==7.7.0
pandas==2.0.3
parsel==1.9.1
parso==0.8.4
pathlib==1.0.1
pdfminer==20191125
pillow==10.2.0
platformdirs==4.0.0
playwright==1.43.0
posthog==3.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.41
Protego==0.3.1
proto-plus==1.23.0
protobuf==4.25.3
prov==2.0.0
psutil==5.9.8
pulsar-client==3.4.0
pure-eval==0.2.2
pyaml==23.12.0
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodome==3.20.0
pydeck==0.8.1b0
PyDispatcher==2.0.7
pydot==2.0.0
pyee==11.1.0
Pygments==2.17.2
pymdown-extensions==10.7.1
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
PyMySQL==0.9.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject==1.3.1
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pysbd==0.3.4
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
pytz==2024.1
pywin32==306
pyxnat==1.6.2
pyzmq==25.1.1
queuelib==1.6.2
ragas==0.1.9
rapidfuzz==3.6.1
rdflib==7.0.0
readability-lxml==0.8.1
referencing==0.34.0
regex==2023.12.25
requests-file==2.0.0
requests-oauthlib==1.4.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
scrapegraphai==0.9.0
Scrapy==2.11.1
selectolax==0.3.21
selenium==4.19.0
sentence-transformers==2.7.0
sentencepiece==0.2.0
sentry-sdk==2.0.1
service-identity==24.1.0
setproctitle==1.3.3
sgmllib3k==1.0.0
shortuuid==1.0.13
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
soupsieve==2.5
sseclient==0.0.27
st-annotated-text==4.0.1
stack-data==0.6.3
starlette==0.36.3
streamlit==1.33.0
streamlit-aggrid==0.3.4
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
striprtf==0.0.26
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
termcolor==2.4.0
threadpoolctl==3.3.0
tiktoken==0.5.2
tinysegmenter==0.3
tld==0.13
tldextract==5.1.2
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchaudio==2.2.1+cu121
torchtext==0.18.0
torchvision==0.17.1+cu121
tornado==6.3.3
tqdm==4.66.3
trafilatura==1.8.1
traitlets==5.13.0
traits==6.3.2
transformers==4.38.2
trio==0.25.0
trio-websocket==0.11.1
trulens-eval==0.27.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typeguard==4.2.1
typer==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
unstructured==0.12.6
unstructured-client==0.18.0
uvicorn==0.28.0
validators==0.28.0
w3lib==2.1.2
wandb==0.16.6
watchdog==4.0.0
watchfiles==0.21.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.
zipp==3.18.1
zope.interface==6.3
